# docs/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Advanced web panel for managing Xray-core servers. Supports VMESS, VLESS, Trojan, Shadowsocks, Hysteria, WireGuard, and more.",
        "title": "3X-UI Panel API",
        "termsOfService": "https://github.com/MHSanaei/3x-ui",
        "contact": {
            "name": "MHSanaei",
            "url": "https://github.com/MHSanaei/3x-ui"
        },
        "license": {
            "name": "GNU GPL v3",
            "url": "https://www.gnu.org/licenses/gpl-3.0.html"
        },
        "version": "3.0.1"
    },
    "host": "localhost:2053",
    "basePath": "/",
    "paths": {
...
        "/panel/setting/getDefaultJsonConfig": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns the default Xray configuration template.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Settings"
                ],
                "summary": "Get default Xray config",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.Msg"
                        }
                    }
                }
            }
        },
...
        "/panel/setting/update": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Updates all panel settings with the provided configuration.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Settings"
                ],
                "summary": "Update settings",
                "parameters": [
                    {
                        "description": "All panel settings",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.AllSetting"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.Msg"
                        }
                    }
                }
            }
        },
...
        "/panel/xray/": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns the Xray configuration template, inbound tags, and outbound test URL.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Xray Settings"
                ],
                "summary": "Get Xray settings",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.Msg"
                        }
                    }
                }
            }
        },
        "/panel/xray/getDefaultJsonConfig": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns the default Xray configuration template.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Xray Settings"
                ],
                "summary": "Get default Xray config",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.Msg"
                        }
                    }
                }
            }
        },
        "/panel/xray/getOutboundsTraffic": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns traffic statistics for all Xray outbounds.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Xray Settings"
                ],
                "summary": "Get outbounds traffic",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.Msg"
                        }
                    }
                }
            }
        },
        "/panel/xray/getXrayResult": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns the current Xray process output/status.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Xray Settings"
                ],
                "summary": "Get Xray result",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.Msg"
                        }
                    }
                }
            }
        },
        "/panel/xray/nord/{action}": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Manages NordVPN integration: list countries, servers, register, set key, get data, delete.",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Xray Settings"
                ],
                "summary": "NordVPN operations",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Action (countries, servers, reg, setKey, data, del)",
                        "name": "action",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Country ID for server list",
                        "name": "countryId",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "NordVPN API token for registration",
                        "name": "token",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Access key",
                        "name": "key",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.Msg"
                        }
                    }
                }
            }
        },
        "/panel/xray/resetOutboundsTraffic": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Resets traffic statistics for a specified outbound by tag.",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Xray Settings"
                ],
                "summary": "Reset outbound traffic",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Outbound tag to reset",
                        "name": "tag",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.Msg"
                        }
                    }
                }
            }
        },
        "/panel/xray/testOutbound": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Tests an outbound configuration and returns the delay/response time. Supports TCP probe mode for parallel-safe testing.",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Xray Settings"
                ],
                "summary": "Test outbound",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Outbound configuration JSON",
                        "name": "outbound",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "JSON array of all outbounds (for dependency resolution)",
                        "name": "allOutbounds",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Test mode: 'tcp' for fast dial-only probe",
                        "name": "mode",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.Msg"
                        }
                    }
                }
            }
        },
        "/panel/xray/update": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Updates the Xray configuration and outbound test URL.",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Xray Settings"
                ],
                "summary": "Update Xray settings",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Xray configuration JSON",
                        "name": "xraySetting",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Outbound test URL",
                        "name": "outboundTestUrl",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.Msg"
                        }
                    }
                }
            }
        },
        "/panel/xray/warp/{action}": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Manages Warp integration: get data, delete, get config, register, set license.",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Xray Settings"
                ],
                "summary": "Warp operations",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Action (data, del, config, reg, license)",
                        "name": "action",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Private key for registration",
                        "name": "privateKey",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Public key for registration",
                        "name": "publicKey",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "License key",
                        "name": "license",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.Msg"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "entity.Msg": {
            "type": "object",
            "properties": {
                "msg": {
                    "description": "Response message text",
                    "type": "string"
                },
                "obj": {
                    "description": "Optional data object"
                },
                "success": {
                    "description": "Indicates if the operation was successful",
                    "type": "boolean"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Bearer token authentication. Use format: \"Bearer \u003ctoken\u003e\". Get your token from Settings → Security → API Token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}